requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
easyocr==1.7.1
pillow==10.3.0
python-multipart==0.0.9
